requirements.txt
accelerate>=0.26.0
bitsandbytes>=0.43.0
datasets>=2.18.0
einops
fvcore
hydra-core>=1.3.0
jinja2
kaggle
loguru
omegaconf>=2.3.0
peft>=0.10.0
pydantic
pytest
python-dotenv
pytorch_lightning
scikit-learn>=1.0.0
sentence-transformers
torch
torchvision
torchaudio
torchmetrics>=0.8.2
transformers>=4.39.0
trl>=0.7.0
wandb>=0.16.0
git+https://github.com/google/gemma_pytorch